Using Weka Models in PHP
In this study, it was determined that there is no interaction between the two amino acids using the user-requested encoding methods such as HP and PP, user-requested encoding techniques such as AAC and AAP, and machine learning methods such as BayesNet, NaiveBayes, RandomForest.
While web technology such as Html, Css, Javascipt, Php is used for this study to work smoothly in web environment, Weka library and Java are used for machine learning part. The PHP / Java Bridge is used to provide a link between these two parts.